updating

+++ TeamProjectMVC/Dto/ProductDTO.cs
@@ -4,7 +4,7 @@
     {
         public string? Name { get; set; }
 
-        public int Quantity { get; set; }
+        public ulong Quantity { get; set; }
 
         public double Price { get; set; }
     }
